Corpus of sample programs (#16)
* Rewrite of results to support add_results, show_results paradigm

* Added program demonstrating UserInputComponent
Modified user_input_component.py to cast value to expected type on init when value is not None
Got rid of get_value from user_input_component.py
Modified raise_type_casting_error's message in CNLError.py

* Renamed input_components.py to non_backend_components.py
Added a text_component stage

* Fixed typo in file_upload_component_template.html

* Each tool gets its own subfolder for uploaded files

* TableSelectorComponents only show user tables unless requested
User created tables cannot begin with __
Check to see if user created tables will overwrite other tables

* BasicBackend tests
Fix column_selector_component.py not overriding paint
Add option for get_column_names_from_table_name to not filter user columns
result_template_of_value doesn't immediately recurse on InputComponents
Fixed typo in select_statement_results.html

* Stop trying to sniff csv file dialects

* Wrote results_demo.py
Added function to allow creation of tables from lists of lists

* Wrote part of roster_aligner.py
Added a result_template_of_list_list to results.py to let programmers display lists of lists as tables

+++ coolNewLanguage/src/component/file_upload_component.py
@@ -1,3 +1,4 @@
+import os.path
 import pathlib
 
 import aiohttp.web_request
@@ -5,7 +6,7 @@ import jinja2
 
 from coolNewLanguage.src import consts
 from coolNewLanguage.src.component.input_component import InputComponent
-from coolNewLanguage.src.stage import config
+from coolNewLanguage.src.stage import config, process
 
 
 class FileUploadComponent(InputComponent):
@@ -30,10 +31,20 @@ class FileUploadComponent(InputComponent):
         if self.value is not None:
             if not isinstance(self.value, aiohttp.web_request.FileField):
                 raise TypeError("Expected value to be an aiohttp FileField")
-            # if FILES_DIR doesn't exist, create it
-            consts.FILES_DIR.mkdir(parents=True, exist_ok=True)
-            # construct path
-            file_path = consts.FILES_DIR.joinpath(self.value.filename)
+            # Check file extension
+            extension = pathlib.Path(self.value.filename).suffix
+            if extension != ('.' + expected_ext):
+                raise ValueError(
+                    f"The uploaded file's extension did not match the expected one. Expected a file with extension "
+                    f"{self.expected_ext}, and instead got {extension}"
+                )
+            # construct path to store file
+            tool = process.running_tool
+            file_path = tool.file_dir.joinpath(self.value.filename)
+            # check that the file doesn't already exist
+            if os.path.isfile(file_path):
+                raise ValueError(f"A file named {self.value.filename} already been uploaded. Please rename the file or "
+                                 f"upload a file with a different name.")
             # write our own copy of the file
             with open(file_path, mode='wb') as f:
                 f.write(self.value.file.read())
